# drymulator/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/command/start": {"post": {"summary": "Start", "operationId": "start_command_start_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "boolean", "title": "Response Start Command Start Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/state/current": {"get": {"summary": "Get Current State", "operationId": "get_current_state_state_current_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatePublic"}}}}}}}, "/state/time": {"get": {"summary": "Get State", "operationId": "get_state_state_time_get", "parameters": [{"name": "second_after", "in": "query", "required": true, "schema": {"type": "integer", "title": "Second After"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatePublic"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"ConfigCreate": {"properties": {"start_time": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Start Time", "default": "2025-03-03T23:48:43.502907"}, "time_speed": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Time Speed", "default": 10.0}}, "type": "object", "title": "ConfigCreate"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "StatePublic": {"properties": {"time_seconds": {"type": "integer", "title": "Time Seconds"}, "fraction_initial": {"type": "number", "title": "Fraction Initial"}, "weight": {"type": "number", "title": "Weight"}}, "type": "object", "required": ["time_seconds", "fraction_initial", "weight"], "title": "StatePublic"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}